+++ swing/common-ui/src/test/java/is/codion/swing/common/ui/icon/IconsTest.java
@@ -21,8 +21,6 @@ package is.codion.swing.common.ui.icon;
 import org.junit.jupiter.api.Test;
 import org.kordamp.ikonli.foundation.Foundation;
 
-import java.awt.Color;
-
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
@@ -36,6 +34,5 @@ public final class IconsTest {
 		assertThrows(IllegalArgumentException.class, () -> icons.add(Foundation.FOUNDATION));
 		assertNotNull(icons.icon(Foundation.ALERT));
 		assertNotNull(icons.icon(Foundation.FOUNDATION));
-		Icons.ICON_COLOR.set(Color.RED);
 	}
 }